Move robot helper classes into subpackage robotcomponents
